requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
pytest==7.4.0